restructured the component directory
