+++ src/App.js
@@ -3,7 +3,6 @@ import Homepage from "./pages/Homepage";
 import About from "./pages/About";
 import Services from "./pages/Services";
 import Connect from "./pages/Connect";
-import NavBar from "./components/NavBar/NavBar";
 
 function App() {
   return (
