+++ src/pages/Homepage.js
@@ -1,16 +1,21 @@
 import { style } from "../Style/Style.js";
-import Logo from "../components/LogoImage/Logo";
-import NavBar from "../components/NavBar/NavBar";
-import HomeBookButton from "../components/Buttons/HomeBookButton.js";
+import Logo from "../components/ReusableComponents/LogoImage/Logo";
+import NavBar from "../components/ReusableComponents/NavBar/NavBar";
+import HomeBookButton from "../components/ReusableComponents/Buttons/HomeBookButton";
 
 export default function Homepage() {
   return (
     <div>
-      <div style={style.heroContainer}>
+      <section style={style.heroContainer}>
         <Logo />
         <HomeBookButton />
         <NavBar />
-      </div>
+      </section>
+      <section>
+        <div>
+          <h1>Testimonials</h1>
+        </div>
+      </section>
     </div>
   );
 }
